requirements.txt
numpy>=1.23.5,<2.1
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx>=1.18.0,<1.20
insightface>=0.7.3
psutil>=5.9.8
tk>=0.1.0
customtkinter>=5.2.2
pillow>=10.0.0
onnxruntime-silicon>=1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu>=1.18.0; sys_platform != 'darwin'
tensorflow>=2.14.0; sys_platform != 'darwin'
opennsfw2>=0.10.2
protobuf>=3.20.0,<5
pygrabber
requests>=2.28.0
tqdm>=4.64.0
pyyaml>=6.0
ffmpeg-python>=0.2.0

# Optional: Lip Sync (for audio-driven mouth movement)
librosa>=0.10.0
soundfile>=0.12.0

# Optional: Face Restoration
gfpgan>=1.3.0
facexlib>=0.3.0

# Optional: Web API
flask>=3.0.0; sys_platform != 'darwin'
fastapi>=0.100.0; sys_platform != 'darwin'
uvicorn>=0.25.0; sys_platform != 'darwin'
pydantic>=2.0.0; sys_platform != 'darwin'
